pyproject.toml
[tool.poetry]
name = "pydev"
version = "0.1.0"
description = ""
authors = ["Nguyen Khac Thanh <nguyenkhacthanh244@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = "^20.8b1"
isort = "^5.8.0"
pylint = "^2.7.4"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
pydev = "pydev.cli:main"
